+++ packages/nimble-components/src/table-column/base/index.ts
@@ -2,9 +2,10 @@ import { attr, nullableNumberConverter } from '@microsoft/fast-element';
 import { FoundationElement } from '@microsoft/fast-foundation';
 import { TableColumnSortDirection } from '../../table/types';
 import {
-    ColumnInternalsOptions,
-    ColumnInternals
+    ColumnInternals,
+    ColumnInternalsOptions
 } from './models/column-internals';
+import type { TableColumnValidity } from './types';
 
 /**
  * The base class for table columns
@@ -12,6 +13,13 @@ import {
 export abstract class TableColumn<
     TColumnConfig = unknown
 > extends FoundationElement {
+    /**
+     * @internal
+     *
+     * Column properties configurable by plugin authors
+     */
+    public readonly columnInternals: ColumnInternals<TColumnConfig> = new ColumnInternals(this.getColumnInternalsOptions());
+
     @attr({ attribute: 'column-id' })
     public columnId?: string;
 
@@ -33,24 +41,15 @@ export abstract class TableColumn<
     @attr({ attribute: 'sorting-disabled', mode: 'boolean' })
     public sortingDisabled = false;
 
-    /**
-     * @internal
-     *
-     * Column properties configurable by plugin authors
-     */
-    public readonly columnInternals: ColumnInternals<TColumnConfig>;
-
-    public constructor(options: ColumnInternalsOptions) {
-        super();
-        if (!options) {
-            throw new Error(
-                'ColumnInternalsOptions must be provided to constructor'
-            );
-        }
-        this.columnInternals = new ColumnInternals(options);
-        this.columnInternals.currentSortDirection = this.sortDirection;
-        this.columnInternals.currentSortIndex = this.sortIndex;
+    public checkValidity(): boolean {
+        return this.columnInternals.validConfiguration;
     }
+
+    public get validity(): TableColumnValidity {
+        return {};
+    }
+
+    protected abstract getColumnInternalsOptions(): ColumnInternalsOptions;
 
     protected sortDirectionChanged(): void {
         if (!this.sortingDisabled) {
@@ -65,15 +64,12 @@ export abstract class TableColumn<
     }
 
     protected sortingDisabledChanged(): void {
-        // Ignore the default value sortingDisabled initialization from undefined to false (which runs before columnInternals is initialized)
-        if (this.columnInternals) {
-            if (this.sortingDisabled) {
-                this.columnInternals.currentSortDirection = TableColumnSortDirection.none;
-                this.columnInternals.currentSortIndex = undefined;
-            } else {
-                this.columnInternals.currentSortDirection = this.sortDirection;
-                this.columnInternals.currentSortIndex = this.sortIndex;
-            }
+        if (this.sortingDisabled) {
+            this.columnInternals.currentSortDirection = TableColumnSortDirection.none;
+            this.columnInternals.currentSortIndex = undefined;
+        } else {
+            this.columnInternals.currentSortDirection = this.sortDirection;
+            this.columnInternals.currentSortIndex = this.sortIndex;
         }
     }
 }