+++ packages/nimble-components/src/table-column/base/tests/table-column-stories-utils.ts
@@ -45,3 +45,8 @@ export const sharedTableArgs = (
         }
     } as const;
 };
+
+export const sharedTableActions = [
+    'selection-change',
+    'column-configuration-change'
+] as const;