+++ packages/nimble-components/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @ni/nimble-components
 
-This log was last generated on Fri, 12 May 2023 23:44:48 GMT and should not be manually modified.
+This log was last generated on Tue, 23 May 2023 16:09:57 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 19.1.1
+
+Tue, 23 May 2023 16:09:57 GMT
+
+### Patches
+
+- Move ColumnInternals to abstract property ([ni/nimble@c03b36b](https://github.com/ni/nimble/commit/c03b36bd9056ef9ad08b156bcaba4d8d2c591c2d))
+
+## 19.1.0
+
+Mon, 22 May 2023 18:39:33 GMT
+
+### Minor changes
+
+- Introduce column configuration validation ([ni/nimble@c23fbe5](https://github.com/ni/nimble/commit/c23fbe55a607508ecc94d7e9d475dc478629cf8a))
+
+## 19.0.2
+
+Wed, 17 May 2023 21:33:40 GMT
+
+### Patches
+
+- Add column-configuration-change event to the nimble-table ([ni/nimble@71351b1](https://github.com/ni/nimble/commit/71351b15e45f68adb8b08db2168b5faf499a49f9))
 
 ## 19.0.1